third_party/blink/web_tests/external/wpt/cookiestore/cookieStore_getAll_arguments.https.any.serviceworker-expected.txt

This is a testharness.js-based test.
[FAIL] cookieStore.getAll with same-origin url path in options
  promise_test: Unhandled rejection with value: object "TypeError: Failed to execute 'getAll' on 'CookieStore': URL must be within Service Worker scope"
[FAIL] cookieStore.getAll with whitespace
  assert_equals: expected 1 but got 0
Harness: the test ran to completion.
